probando con el file_get

+++ src/Service/WapiRequest/ProcessCMD.php
@@ -39,12 +39,15 @@ class ProcessCMD {
         file_put_contents('exit3_'.$waId.'_'.'.txt', $filename);
         $filename = $filename.'.json';
         file_put_contents('exit4_'.$waId.'_'.'.txt', $filename);
-        $exist = ($this->filesystem->exists($filename)) ? true : false;
+        $exist = file_get_contents($filename);
+        if($exist === false) {
+            return false;
+        }
         file_put_contents('exit5_'.$exist.'_'.'.json', json_encode([
             'exit' => $exist, 'file' => $filename
         ]));
 
-        return $exist;
+        return true;
     }
 
     /**
